Move extension function out of the InstallmentModel into InstallmentModelExtensions
