+++ card/src/main/java/com/adyen/checkout/card/internal/ui/model/InstallmentModel.kt
@@ -8,12 +8,7 @@
 
 package com.adyen.checkout.card.internal.ui.model
 
-import androidx.compose.runtime.Composable
-import com.adyen.checkout.core.common.internal.helper.LocalLocale
-import com.adyen.checkout.core.common.localization.CheckoutLocalizationKey
-import com.adyen.checkout.core.common.localization.internal.helper.resolveString
 import com.adyen.checkout.core.components.data.model.Amount
-import com.adyen.checkout.core.components.data.model.format
 
 internal sealed interface InstallmentModel {
     data object OneTime : InstallmentModel
@@ -23,27 +18,4 @@ internal sealed interface InstallmentModel {
         val amountPerInstallment: Amount?,
         val showAmount: Boolean,
     ) : InstallmentModel
-}
-
-@Composable
-internal fun InstallmentModel.toDisplayText(): String {
-    return when (this) {
-        InstallmentModel.OneTime ->
-            resolveString(CheckoutLocalizationKey.CARD_INSTALLMENTS_ONE_TIME)
-
-        InstallmentModel.Revolving ->
-            resolveString(CheckoutLocalizationKey.CARD_INSTALLMENTS_REVOLVING)
-
-        is InstallmentModel.Regular -> {
-            if (showAmount && amountPerInstallment != null) {
-                resolveString(
-                    CheckoutLocalizationKey.CARD_INSTALLMENTS_REGULAR_WITH_PRICE,
-                    numberOfInstallments,
-                    amountPerInstallment.format(LocalLocale.current),
-                )
-            } else {
-                resolveString(CheckoutLocalizationKey.CARD_INSTALLMENTS_REGULAR, numberOfInstallments)
-            }
-        }
-    }
 }
